// frontend/src/api/endpoints/generated.ts
export interface SubClassEntry {
  id?: string;
  name?: string;
  description?: string;
}

export interface FeatureDef {
  level?: number;
  id?: string;
  name?: string;
}

export type SpellcastingEntrySpellSlots = {[key: string]: number};

export type SpellcastingEntrySpellLists = {[key: string]: string[]};

export interface SpellcastingEntry {
  type?: string;
  ability?: string;
  cantripsKnown?: number[];
  preparedSpells?: number[];
  spellsKnown?: number[];
  spellSlots?: SpellcastingEntrySpellSlots;
  ritualCasting?: boolean;
  spellLists?: SpellcastingEntrySpellLists;
}

export interface ClassEntry {
  id?: string;
  name?: string;
  hitDie?: string;
  savingThrows?: string[];
  primaryAbility?: string[];
  skillChoices?: number;
  skillPool?: string[];
  spellcaster?: boolean;
  subClasses?: SubClassEntry[];
  subclassLevel?: number;
  features?: FeatureDef[];
  spellcasting?: SpellcastingEntry;
}